ai_agents/audio_transcript/pyproject.toml
[project]
name = "audio-transcript"
version = "0.1.0"
description = ""
authors = [
    {name = "WandersonCorreia",email = "wandersoncorreia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydub (>=0.25.1,<0.26.0)",
    "openai (>=1.109.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.50.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "audio_transcript", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
